working to fix theme
