--- app/(tabs)/index.tsx
@@ -1,31 +1,0 @@
-import React from 'react';
-import { StyleSheet, View, Text } from 'react-native';
-import { ActivityIndicator } from 'react-native-paper';
-
-export default function TabOneScreen() {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Home</Text>
-      <View style={styles.separator} />
-      <ActivityIndicator />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-  separator: {
-    marginVertical: 30,
-    borderBottomWidth: 1,
-    borderBottomColor: 'rgba(255, 255, 255, 0.1)',
-    width: '80%',
-  },
-});
